Removing thunks to get everything working
