+++ src/actions.js
@@ -1,90 +1,19 @@
 export const types = {
-  CREATE_ELEMENT: 'CREATE_ELEMENT',
   SET_ATTRIBUTE: 'SET_ATTRIBUTE',
   REMOVE_ATTRIBUTE: 'REMOVE_ATTRIBUTE',
-  APPEND_CHILD: 'APPEND_CHILD',
   INSERT_CHILD: 'INSERT_CHILD',
   REPLACE_CHILD: 'REPLACE_CHILD',
   REMOVE_CHILD: 'REMOVE_CHILD',
   UPDATE_CHILD: 'UPDATE_CHILD',
-  INSERT_BEFORE: 'INSERT_BEFORE',
-  CREATE_THUNK: 'CREATE_THUNK',
-  UPDATE_THUNK: 'UPDATE_THUNK',
-  DESTROY_THUNK: 'DESTROY_THUNK'
+  INSERT_BEFORE: 'INSERT_BEFORE'
 }
 
-export function createElement (vnode) {
+export function setAttribute (name, value, previousValue) {
   return {
-    type: types.CREATE_ELEMENT,
-    vnode
-  }
-}
-
-export function updateChild (vnode, index, actions) {
-  return {
-    type: types.UPDATE_CHILD,
-    vnode,
-    index,
-    actions
-  }
-}
-
-export function insertChild (vnode, path, index) {
-  return {
-    type: types.INSERT_CHILD,
-    vnode,
-    path,
-    index
-  }
-}
-
-export function removeChild (vnode, index) {
-  return {
-    type: types.REMOVE_CHILD,
-    vnode,
-    index
-  }
-}
-
-export function replaceChild (previous, next, path, index) {
-  return {
-    type: types.REPLACE_CHILD,
-    previous,
-    next,
-    path,
-    index
-  }
-}
-
-export function appendChild (vnode) {
-  return {
-    type: types.APPEND_CHILD,
-    vnode
-  }
-}
-
-export function createThunk (thunk, path) {
-  return {
-    type: types.CREATE_THUNK,
-    thunk,
-    path
-  }
-}
-
-export function updateThunk (thunk, previousThunk, path, index) {
-  return {
-    type: types.UPDATE_THUNK,
-    thunk,
-    previousThunk,
-    path,
-    index
-  }
-}
-
-export function destroyThunk (thunk) {
-  return {
-    type: types.DESTROY_THUNK,
-    thunk
+    type: types.SET_ATTRIBUTE,
+    name,
+    value,
+    previousValue
   }
 }
 
@@ -97,19 +26,44 @@ export function removeAttribute (name, previousValue) {
   }
 }
 
-export function setAttribute (name, value, previousValue) {
+export function insertChild (vnode, index) {
   return {
-    type: types.SET_ATTRIBUTE,
-    name,
-    value,
-    previousValue
+    type: types.INSERT_CHILD,
+    vnode,
+    index
   }
 }
 
-export function insertBefore (vnode, position) {
+export function replaceChild (previous, next, index) {
+  return {
+    type: types.REPLACE_CHILD,
+    previous,
+    next,
+    index
+  }
+}
+
+export function removeChild (vnode, index) {
+  return {
+    type: types.REMOVE_CHILD,
+    vnode,
+    index
+  }
+}
+
+export function updateChild (vnode, index, actions) {
+  return {
+    type: types.UPDATE_CHILD,
+    vnode,
+    index,
+    actions
+  }
+}
+
+export function insertBefore (vnode, index) {
   return {
     type: types.INSERT_BEFORE,
     vnode,
-    position
+    index
   }
 }
