+++ test/_diffNode.js
@@ -1,0 +1,4 @@
+import test from 'tape'
+import {diffNode} from '../src/diff'
+import * as actions from '../src/actions'
+import h from '../src/element'
