+++ test/_renderString.js
@@ -1,0 +1,4 @@
+import test from 'tape'
+import renderString from '../src/renderString'
+import * as actions from '../src/actions'
+import h from '../src/element'
